extract buttons component
