+++ src/components/ArchitectureBlock/TreeItem.tsx
@@ -2,6 +2,7 @@ import React, { useState, Dispatch, ChangeEvent, MouseEvent } from 'react';
 import toast from 'react-hot-toast';
 import { validateFileName } from '../../utils/validateInput';
 import { structureStateType } from '../../helpers/types';
+import { ButtonEditFilename, ButtonRemoveFile } from '../Buttons';
 
 export const TreeItem = ({
   structure,
@@ -10,7 +11,6 @@ export const TreeItem = ({
   name,
   isFolder,
   ancestor,
-  path,
 }: {
   structure: structureStateType;
   dispatchStructure: Dispatch<any>;
@@ -18,7 +18,6 @@ export const TreeItem = ({
   name: string;
   isFolder: boolean;
   ancestor: string;
-  path: string;
 }) => {
   const [visible, setVisible] = useState(false);
   const [isEdit, setIsEdit] = useState(false);
@@ -37,16 +36,6 @@ export const TreeItem = ({
 
   const hideButtons = () => {
     if (name !== 'src' && name !== 'App') setVisible(false);
-  };
-
-  const removeItem = (e: MouseEvent<HTMLButtonElement>) => {
-    e.stopPropagation();
-    dispatchStructure({ type: 'REMOVE', payload: { id: id } });
-  };
-
-  const editNameItem = (e: MouseEvent<HTMLButtonElement>) => {
-    e.stopPropagation();
-    setIsEdit(true);
   };
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
@@ -164,46 +153,9 @@ export const TreeItem = ({
               </svg>
             </button>
           ) : (
-            <button onClick={editNameItem}>
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="icon icon-tabler icon-tabler-pencil transform hover:rotate-45 transition duration-200"
-                width="20"
-                height="20"
-                viewBox="0 0 24 24"
-                strokeWidth="1.5"
-                stroke="#3d3d3d"
-                fill="none"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              >
-                <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-                <path d="M4 20h4l10.5 -10.5a1.5 1.5 0 0 0 -4 -4l-10.5 10.5v4" />
-                <line x1="13.5" y1="6.5" x2="17.5" y2="10.5" />
-              </svg>
-            </button>
+            <ButtonEditFilename setIsEdit={setIsEdit} />
           )}
-          <button onClick={removeItem}>
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              className="icon icon-tabler icon-tabler-trash transform hover:rotate-45 transition duration-200"
-              width="20"
-              height="20"
-              viewBox="0 0 24 24"
-              strokeWidth="1.5"
-              stroke="#3d3d3d"
-              fill="none"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            >
-              <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-              <line x1="4" y1="7" x2="20" y2="7" />
-              <line x1="10" y1="11" x2="10" y2="17" />
-              <line x1="14" y1="11" x2="14" y2="17" />
-              <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" />
-              <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" />
-            </svg>
-          </button>
+          <ButtonRemoveFile id={id} dispatchStructure={dispatchStructure} />
         </div>
       </div>
       <ul className={`${display ? 'visible' : 'hidden'} ml-4 border-l-2 border-gray-200`}>
@@ -215,7 +167,6 @@ export const TreeItem = ({
             name={ele.name}
             isFolder={ele.isFolder}
             ancestor={ele.ancestor}
-            path={ele.path}
             dispatchStructure={dispatchStructure}
           />
         ))}
