+++ src/components/Buttons/ButtonAddScript.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 export const ButtonAddScript = () => {
   return (
     <button
-      className="my-1 mx-auto bg-green-500 opacity-100 px-4 py-1 outline-none
+      className="my-1 mx-auto bg-gray-900 opacity-100 px-4 py-1 outline-none
                               tracking-wider text-white rounded hover:opacity-90 focus:outline-none transition duration-250"
     >
       Add
